Change term terminator to comma
Parsing structs made it apparent that the original grammar required two
term terminators (comma inside a sequence, semicolon elsewhere).  Using
two different terminators complicates the parser.  It also requires
users to remember which character is the proper terminator in which
context.

I could have standardized on either comma or semicolon as the
terminator.  I chose comma because the terminator inside a sequence is
likely to exist explicitly in Amalog source code.  The terminator inside
a database is likely to be inserted by the scanner so nobody will ever
see it.  Since nobody will see it, it doesn't matter much what it is.
The comma is definitely the standard for separating arguments in a
function call.  Since sequences, as part of a struct, will fill that
syntactic role it seems wise to choose comma as the separator.

As an added bonus, this almost allows one to copy-paste Prolog code into
Amalog because Prolog effectively uses comma as a term terminator inside
clause bodies.

One test still fails.  That's because formatting always inserts newlines
but that's not wanted when formatting terms inside a sequence.