Comma insertion to close a sequence
Following the lead of BCPL, have the scanner insert a trailing comma
right before the closing parenthesis of a sequence if one is missing.
The rule for inserting a comma is easy to understand.  Knowing that the
term terminator will always be present allows us to keep the parser
simple.