Move tests to file system
All the old tests were hard coded into the Go test harness.  The Amalog
code used in those tests will be of general interest to any Amalog
implementation.  By storing the Amalog test code in the file system and
having Go's test harness read it from there, we take a step towards a
generic Amalog test suite.

Eventually the tests/ directory will be split off to a separate
repository which implementations can include as a submodule.  While the
test suite is rapidly evolving it's easier to leave it here.