First (failing) TAP test
The tests under tests/tap/** are Amalog scripts which generate output
for the Test Anything Protocol.  This gives us an executable test suite
which can be shared across all language implementation.  TAP is already
implemented in many languages and easy to implement where it's not
available.  The 'prove' command which runs TAP tests is flexible enough
to be the test harness for any language and provides helpful
infrastructure for running and view tests.