Implement module search path
When importing Amalog modules, we search inside any "amalog_modules"
directory contained within a parent of the directory that contains the
source code or the executable.

This allows users to have Amalog modules that are specific to a project
or a user or the entire system.  The most specific directory takes
precedence.

Many languages choose this search path based on an environment
variable.  In my experience that leads to a lot of confusion and
requires extra work to configure the environment.  One has to create a
directory structure for the module source code anyway, so we might as
well establish a convention and rely on that.