Sketch architecture for loading dependencies
I want to make sure that the semantics of modules allow for parallel
loading.  The best way I can think of to guarantee that is to implement
a parallel loader.

This is a basic graph traversal algorithm performed with N worker
threads.  Each module is loaded only once.  Cycles are allowed.