Parts of standard library needed for Hello World
I created modules for amalog.org/std/{io,env} which should be all that's
needed to get the Hello World test to pass.  The TODO comments indicate
Amalog predicates that need to be implemented.

The next step is probably to implement Machine.Call() in Go.  It should
be able to invoke main/0 in the script to trigger execution of all
predicates downstream.