Remove all references to the "world" argument
My original thought was for the world argument to be a proxy between an
Amalog program and the outside world.  As I thought through all the
details, I realized that it wouldn't work as I had envisioned it.  Most
of the use cases I had imagined can be solved with static analysis so I
don't think we lose anything by removing the "world".